requirements.txt
# Forge Requirements Assistant - Dependencies

# Core dependencies
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Frontend
streamlit>=1.40.0
audio-recorder-streamlit>=0.0.8
openai>=1.0.0

# Voice Agent Server
fastapi>=0.115.0
uvicorn>=0.32.0
websockets>=13.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
